backend/requirements.txt
# Lawsarthi Backend Dependencies

# Flask Web Framework
flask==3.0.0
flask-cors==4.0.0

# PostgreSQL with pgvector
psycopg2-binary==2.9.9

# Machine Learning - Sentence Embeddings
sentence-transformers==2.2.2
torch>=2.0.0

# Utilities
python-dotenv==1.0.0
